Add a structure and query for reporting descriptor accesses

+++ renderdoc/data/glsl_shaders.h
@@ -41,7 +41,7 @@ rdcstr InsertSnippetAfterVersion(ShaderType type, const char *source, int len, c
 // for unit tests
 struct ShaderReflection;
 struct ShaderBindpointMapping;
-enum class ShaderStage : uint32_t;
+enum class ShaderStage : uint8_t;
 using ReflectionMaker =
     std::function<void(ShaderStage stage, const rdcstr &source, const rdcstr &entryPoint,
                        ShaderReflection &refl, ShaderBindpointMapping &mapping)>;
